Update URI_1176.java
it won't work with an integer array since the question specified that, "All the Fibonacci numbers calculated in this program must fit in an unsigned 64 bits number."

+++ URI_1176.java
@@ -15,7 +15,7 @@ public class URI_1176 {
 
 	public static void main(String[] args) {
 		int T, N, i, j;
-		int fibo[]=new int[61];
+		long fibo[]=new long[61];
 		fibo[0] = 0;
 		fibo[1] = 1;
